Add parser and class for pre-requirements file
Pre-requirements file is the new way of configuring Prequ.  It controls
how the `requirements*.txt` files are generated.  It should be named
`requirements.pre` and placed in the project root directory.  It
supersedes the `requirements*.in` files.

Add class `PreRequirements` for defining the contents of the
pre-requirements file and code for parsing them.  Includes some tests
too.