Use setup.cfg as default pre-requirements file
Add basic INI parser for parsing [prequ] section from setup.cfg and use
it as the default pre-requirements file.

Update README to state the new default way of configuration.

The requirements.pre file is still supported, but setup.cfg is
preferred: requirements.pre file is not read if there is a [prequ]
section in setup.cfg.