Demand wheel lines to use == version specifier
It doesn't make sense to use anything but the equality operator in the
version specifier of requirements which are built as wheels, since the
wheels are built for just the single version.

This also makes it simpler to implement "prequ check" command for
checking if generated requirements files are up-to-date and all wheels
are built.