Do not allow conflicting pinned versions
The implementation of get_pinned_version (and its main user,
is_pinned_requirement) didn't work correctly when there were conflicting
specifiers: It used to ignore all of the non-pinning specifiers and that
caused a specifier like "==1.8,>1.9" to return a pinned version of "1.8"
even though that is a conflicting specifier (and shouldn't return any
matching versions).

Fix this by filtering the possibly mathcing pinned version with the
`filter` method of the specifier so that non-matching versions are not
returned.

This also removes the need for the ImpossibleConstraint exception, so
remove it.

This amends the previous commit f456185.