Merge setup.cfg with requirements*.in
Make the PrequConfiguration.from_directory construct the configuration
so that options and source requirement sets from the setup.cfg are
merged with source requirement sets defined by the in files.

This makes it possible to have some of the source requirement sets
defined in the setup.cfg and still have e.g. `requirements-local.in` for
additional set of source requirements.