Detect annotation and other options from output files
Make the three configuration options "annotate", "header" and
"generate_hashes" to allow a third possible value "auto" in addition to
previously allowed "yes" or "no".  The "auto" value is also the new
default.  When any of these options is configured as "auto", then its
value is determined by reading all of the already generated output files
(requirements*.txt) and checking if it was enabled in the previous run.
If no output files are generated yet, then following default values are
used:

  - annotate: no
  - header: yes
  - generate_hashes: no