Make "check" work with conf option detection
Amend (and simplify) the "prequ check" command so that it will work
again now that "annotate", "header" and "generate_hashes" are detected
from the output files:

Rather than adding a constraint line "-c requirements-X.txt" to the
corresponding requirements in-file content, just initialize the
corresponding temporary requirements txt file with the contents of the
real txt file.

Also add tests for "prequ check" do some refactoring to the test code so
that this is easier: Move pip_conf fixtures to conftest and move CLI
command invocation and configuration creation code to tests.utils.