Fix "prequ check --silent" to really be silent
The "--silent" option to "prequ check" command should make it to not
generate any output, except in errors.  A requirement*.txt file being
outdated is not an error, but just something that should be reported via
the exit code when being silent.

Make the check command not output "requirement*.txt is outdated" when
ran with the silent flag, but rather just set the non-zero exit code.