Replace fileutils with file_replacer
The AtomicSaver class was the only used feature from fileutils. It was a
bit too complex to my taste and wasn't well enough tested, so create a
new simpler class called `FileReplacer` for similar functionality and
use that instead.

Remove the fileutils module as it is not used anymore.

Add tests for the new module.