Fix for pinned requirements with a period in name
There was a bug in LocalRequirementsRepository (LRR) that caused a name
of a requirement to render invalid if there was a period in its name and
it was already pinned (i.e. exists in the compiled output file), namely
the period was converted to dash.  E.g. "zope.interface" was converted
to "zope-interface".

This bug was caused by the name normalization improvements done in
commit 6934c263, because LRR used to use the normalized key of the
requirement as the name of its returned InstallRequirement when the
requirement was pinned.

Fix the bug by changing LRR to use the name of the existing pinned
requirement rather than the normalized key.

Fixes #21