Fix a few requirement key vs name usages
Name of the distribution may contain a period which is normalized to a
dash in the key.  If distribution package has a period in its name, it's
not installable when the period is replaced with a dash.  Therefore we
must be extra careful to not mix up keys and names.

Fix each usage of requirement key to a name where appropriate.