Tox,AppVeyor: Test fewer Pip versions
There is no need to test two different Pip 8 versions.  The 8.0.x is
most demanding for us, so use it as the Pip 8 version to test.

Remove the piplatest, since it is currently the same as Pip 9 and when
Pip 10 will be released, it will most likely break, so there is no point
even testing with it.