PyPIRepository: Remove unneeded build dir clean-up code
This code used to move the temporary build directory to a new location
under self.build_dir so that now trash is left behind.  That is now
unneeded, because the `reqset.cleanup_files()` call below removes the
build directory anyway.

This also makes it work with Pip 10, since it seems that the
`InstallRequirement._correct_build_location` method is broken in Pip 10.
(It does `assert self._ideal_build_dir.path` with `_ideal_build_dir`
being a string.)