Fix Pip 10 dependency resolving
The dependency calculation in the PyPIRepository resolver was
implemented incorrectly.  It resolved all deps of the given ireq deeply
with Pip, but it should have resolved only one round so that we can use
our resolver for the result to check conflicts/constraints.

This also makes the code in the Pip 10 branch more similar to the code
in Pip 9 branch.