Improve resolver failure error messages
Collect the logs yielded by Pip when dependency resolution fails in
PyPIRepository._get_dependencies and raise an exception that has those
log messages attached so that they can be shown to the user.

This is especially important if user has no setuptools installed,
because then it is very common to get an error about missing setuptools
when calculating dependencies of a package which is available only a
sdist and Pip tries to invoke `python setup.py egg_info` for it.  The
reason for that error was totally invisible before this change.