Fix unsafe package reporting
Unsafe dependencies of packages should generate a section to the output
file with header:

    The following packages are considered to be unsafe in a requirements
    file

followed by the list of collecetd unsafe dependencies.  This feature
exists, but was broken in commit debd8d5c0.

Make this feature work again and update tests so that they check also
the resulting unsafe packages of the resolver.