Tox: Make sure master version of Pip gets installed
Now that we have an upper bound for Pip version in the Prequ
install_requires list, it seems that when Tox installs the Prequ
package, the Pip will be downgraded if it's newer than the specified
upper bound version.  Make sure that the master version is actually used
by installing the Pip from the master URL in a pre-command instead.

Also change the URL from a Git clone to a tar.gz archive, because that
is faster to install.